requirements.txt

streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
yfinance>=0.2.0
pandas_ta>=0.3.14b0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
# AI/ML Stack
torch>=2.0.0
tensorflow>=2.14.0
stable-baselines3>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
pyportfolioopt>=1.5.5
# DeFi & Blockchain
web3>=6.11.0
solana>=0.30.0
eth-account>=0.9.0
# Infrastructure & Utils
redis>=5.0.1
celery>=5.3.4
cryptography>=41.0.5
twilio>=8.5.0
discord.py>=2.3.2
python-telegram-bot>=20.6
matplotlib>=3.8.0
seaborn>=0.13.0
backtrader>=1.9.78.123
quantlib>=1.32
simpy>=4.0.2
gymnasium>=0.29.1
